- fixed CRITICAL bug that caused workdir being deleted even when it is not a temporary directory

+++ modules/db_scen_remove.R
@@ -91,7 +91,9 @@ closeScenario <- function(){
   }
   resetWidgetsOnClose <<- TRUE
   
-  unlink(list.files(workDir, recursive = TRUE))
+  if(useTempDir){
+    unlink(list.files(workDir, recursive = TRUE))
+  }
   if(is.R6(activeScen))
     flog.debug("Scenario: '%s' closed.", activeScen$getScenName())
   # reset input data
